+++ spring-security-course/src/main/java/com/example/demo/jwt/JwtUsernamePasswordAuthenticationFilter.java
@@ -1,6 +1,8 @@
 package com.example.demo.jwt;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -8,9 +10,12 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
+import javax.servlet.FilterChain;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.time.LocalDate;
+import java.util.Date;
 
 public class JwtUsernamePasswordAuthenticationFilter extends UsernamePasswordAuthenticationFilter {
 
@@ -28,9 +33,26 @@ public class JwtUsernamePasswordAuthenticationFilter extends UsernamePasswordAut
                     .readValue(request.getInputStream(), UsernamePasswordAuthenticationRequest.class);
             final Authentication authentication = new UsernamePasswordAuthenticationToken(
                     authenticationRequest.getUsername(), authenticationRequest.getPassword());
-            return authenticationManager.authenticate(authentication);
+            final Authentication authenticate = authenticationManager.authenticate(authentication);
+            return authenticate;
         } catch (IOException e) {
             throw new BadCredentialsException("Error reading request", e);
         }
+    }
+
+    @Override
+    protected void successfulAuthentication(HttpServletRequest request,
+                                            HttpServletResponse response,
+                                            FilterChain chain,
+                                            Authentication authResult) {
+        final String key = "secure-secure-secure-secure-secure-secure-secure-secure-secure-secure";
+        final String token = Jwts.builder()
+                .setSubject(authResult.getName())
+                .claim("authorities", authResult.getAuthorities())
+                .setIssuedAt(new Date())
+                .setExpiration(java.sql.Date.valueOf(LocalDate.now().plusWeeks(2)))
+                .signWith(Keys.hmacShaKeyFor(key.getBytes()))
+                .compact();
+        response.addHeader("Authorization", "Bearer " + token);
     }
 }