+++ spring-security-course/History.md
@@ -19,4 +19,4 @@ History of implementation
 16. Implement [Remember Me](Howto.md#10---remember-me)
 17. Implement [Logout](Howto.md#11---logout)
 18. Implement [DB Authentication](Howto.md#11---db-authentication)
-19. [Start](Howto.md#12--json-web-token-i) JSON Web Token (JWT). 
+19. Use [JSON Web Token (JWT)](Howto.md#12--json-web-token-i).